Per-layer independent eviction via DynamicCache slicing, ready for GPU training
Core changes:
- Replace hook-based eviction with DynamicCache slicing (per_layer_generate.py):
  fresh prefill → index_select per layer → manual greedy decode with true position_ids
- env.py: 28 agents (one per layer), resident_mask [L,T], terminal reward uses
  generate_with_per_layer_eviction; blended correctness + soft attention alignment
- capture.py: .float().cpu() on K/V to fix bfloat16→numpy crash on CUDA server
- eval.py: --budget CLI arg (was hardcoded), score_with_resident uses cache slicing

Cleanup:
- Delete dead files: phase2.yaml, auc.py, correctness.py, vendor/apply_eviction.py
- Rename phase1.yaml → train.yaml (device: auto, save_path: checkpoints/run)
- train.py: remove unused ActionMasker import and mask_fn; fix gamma/gae_lambda
  defaults to 1.0 (Monte Carlo); update default --config to train.yaml
- max_new_tokens: 512 → 524 to match internal-signals repo
- tests/test_fixes.py: update phase1 → train config references

Diagnostics:
- diagnostics/per_head_importance_correlation.py: empirically validates per-layer
  granularity choice (Spearman ρ=0.67 between KV-heads using actual attention
  weights - borderline, 28 agents justified over 56 given sparse reward)

+++ kv-eviction-gym/src/kv_gym/capture.py
@@ -74,8 +74,8 @@ def capture(
     K_list, V_list = [], []
     for l in range(n_layers):
         k, v = _get_kv(l)
-        K_list.append(k.squeeze(0).cpu())  # [n_kv_heads, T, D]
-        V_list.append(v.squeeze(0).cpu())
+        K_list.append(k.squeeze(0).float().cpu())  # [n_kv_heads, T, D]; float32 for numpy compat
+        V_list.append(v.squeeze(0).float().cpu())
 
     return AllHeadCapture(
         K=torch.stack(K_list),            # [L, n_kv_heads, T, D]
